requirements.txt
absl-py==0.15.0
anyio==3.5.0
asgiref==3.5.0
astunparse==1.6.3
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.2
cycler==0.11.0
fastapi==0.75.2
flatbuffers==1.12
fonttools==4.32.0
gast==0.4.0
google-auth==2.9.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.34.1
h11==0.13.0
h5py==3.1.0
idna==3.3
importlib-metadata==4.12.0
joblib==1.1.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.2
numpy==1.19.5
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.3
Pillow==9.1.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.9.1
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.1.1
scipy==1.9.0
seaborn==0.11.2
six==1.15.0
sklearn==0.0
sniffio==1.2.0
starlette==0.17.1
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
threadpoolctl==3.1.0
typing-extensions==3.7.4.3
urllib3==1.26.11
uvicorn==0.17.6
Werkzeug==2.2.1
wrapt==1.12.1
zipp==3.8.1
